+++ src/main/java/com/herokuapp/domain/khachhang/KhachHangDomain.java
@@ -7,7 +7,7 @@ public class KhachHangDomain extends AbstractsDomain<Khachhang> {
 	private String makh;
 	private String diachi;
 	private String ho;
-	private int sdt;
+	private Long sdt;
 	private String ten;
 
 	public KhachHangDomain() {
@@ -37,11 +37,11 @@ public class KhachHangDomain extends AbstractsDomain<Khachhang> {
 		this.ho = ho;
 	}
 
-	public int getSdt() {
+	public Long getSdt() {
 		return sdt;
 	}
 
-	public void setSdt(int sdt) {
+	public void setSdt(Long sdt) {
 		this.sdt = sdt;
 	}
